refactor(api): remove initialize database
