+++ api/src/controllers/index.js
@@ -1,34 +1,46 @@
-const getEvents = require('./event/getEvents')
 const { signupUser, loginUser, protectedUser, } = require('./user')
+
+const getEvents = require('./event/getEvents')
 const { createEvent } = require('./event/createEvent')
+const createMasiveEvents = require('./event/createMasiveEvents')
+const setTopEvent = require('./event/setTopEvent')
+
 const getCategories = require('./category/getCategories')
 const { getEventsByCategory } = require('./category/getEventsByCategory')
+const setCategories = require('./category/setCategories')
+const createCategory = require('./category/createCategory')
+
 const setCountries = require('./location/setCountries')
 const getCountries = require('./location/getCountries')
 const setStates = require('./location/setStates')
 const getStates = require('./location/getStatesByCounty')
 const setCities = require('./location/setCities')
 const getCities = require('./location/getCitiesByState')
-const createCategory = require('./category/createCategory')
-const setTopEvent = require('./event/setTopEvent')
-const { getSales, postSales} = require('./getSales')
+
+const { getSales, postSales } = require('./getSales')
 
 module.exports = {
-    getEvents,
-    createEvent,
-    getCategories,
     signupUser,
     loginUser,
     protectedUser,
+
+    getEvents,
+    createEvent,
+    createMasiveEvents,
+    setTopEvent,
+
+    getCategories,
     getEventsByCategory,
+    createCategory,
+    setCategories,
+
     setCountries,
     getCountries,
     setStates,
     getStates,
     setCities,
     getCities,
-    createCategory,
-    setTopEvent,
+
     getSales,
-    postSales
+    postSales,
 }
