+++ api/src/db.js
@@ -1,17 +1,16 @@
-require('dotenv').config(); 
+require('dotenv').config();
 const fs = require('fs');
 const path = require('path');
 const { Sequelize } = require('sequelize');
 const {
   DB_USER, DB_PASSWORD, DB_HOST, DB_NAME
 } = process.env;
-const {getEvents, getTaxonomies} = require('./utils/loadDb')
 
 const sequelize = new Sequelize(`postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}/${DB_NAME}`, {
-    logging: false,
-    native: false,
-  });
-  
+  logging: false,
+  native: false,
+});
+
 
 const basename = path.basename(__filename);
 
@@ -59,13 +58,7 @@ Ticket.belongsTo(Event)
 States.belongsTo(Countries, { foreignKey: 'country_id' });
 Cities.belongsTo(States, { foreignKey: 'state_id' });
 
-const initializeDatabase = async () => {
-  await getTaxonomies(Category); // Cargar categorías
-  await getEvents(Event); // Cargar eventos
-}
-
 module.exports = {
   ...sequelize.models, // para poder importar los modelos así: const { Product, User } = require('./db.js');
-  conn: sequelize,
-  initializeDatabase // para importart la conexión { conn } = require('./db.js');
+  conn: sequelize, // para importart la conexión { conn } = require('./db.js');
 };
