+++ api/src/routes/categories.js
@@ -1,9 +1,10 @@
 const { Router } = require("express")
 const router = Router()
-const { getCategories, getEventsByCategory, createCategory } = require('../controllers/index')
+const { getCategories, getEventsByCategory, createCategory, setCategories } = require('../controllers/index')
 
 router.get('/categories', getCategories)
 router.post('/categories', createCategory)
+router.put('/categories/masive', setCategories)
 
 router.get('/events-by-category', getEventsByCategory)
