+++ api/src/routes/events.js
@@ -1,14 +1,15 @@
 const { Router } = require("express")
 const router = Router()
-const { getEvents, createEvent, setTopEvent } = require('../controllers/index')
+const { getEvents, createEvent, setTopEvent, createMasiveEvents } = require('../controllers/index')
 const multer = require('multer')
 const storage = multer.memoryStorage();
 const upload = multer({ storage })
 
 router.get('/events', getEvents)
 
-router.post('/events', upload.single("event_image"), createEvent) //definir si misma ruta u otra
+router.post('/events', upload.single("event_image"), createEvent)
 
 router.patch('/events/:id', setTopEvent)
+router.put('/events/masive', createMasiveEvents)
 
 module.exports = router;
